+++ src/entity.rs
@@ -1,26 +1,36 @@
 use std::borrow::Cow;
 
-use crate::error::Error;
+use crate::error::ParseError;
 use crate::output::Normalizer;
+use crate::Span;
 
-pub(crate) fn parse_text(content: Cow<str>) -> Result<Cow<str>, Error> {
-    parse_content(content, false)
+pub(crate) fn parse_text(content: Cow<str>, base_position: usize) -> Result<Cow<str>, ParseError> {
+    parse_content(content, false, base_position)
 }
 
-pub(crate) fn parse_attribute(content: Cow<str>) -> Result<Cow<str>, Error> {
-    parse_content(content, true)
+pub(crate) fn parse_attribute(
+    content: Cow<str>,
+    base_position: usize,
+) -> Result<Cow<str>, ParseError> {
+    parse_content(content, true, base_position)
 }
 
-fn parse_content(content: Cow<str>, attribute: bool) -> Result<Cow<str>, Error> {
+fn parse_content(
+    content: Cow<str>,
+    attribute: bool,
+    base_position: usize,
+) -> Result<Cow<str>, ParseError> {
     let mut result = String::new();
-    let mut chars = content.chars().peekable();
+    let mut chars = content.char_indices().peekable();
     let mut change = false;
-    while let Some(c) = chars.next() {
+    while let Some((position, c)) = chars.next() {
         // https://www.w3.org/TR/xml/#sec-line-ends
         if c == '\r' {
-            if chars.peek() == Some(&'\n') {
-                // consume next char
-                chars.next();
+            if let Some((_, peeked)) = chars.peek() {
+                if peeked == &'\n' {
+                    // consume next char
+                    chars.next();
+                }
             }
             if !attribute {
                 result.push('\n');
@@ -32,31 +42,44 @@ fn parse_content(content: Cow<str>, attribute: bool) -> Result<Cow<str>, Error> 
         } else if c == '&' {
             let mut entity = String::new();
             let mut is_complete = false;
-            for c in chars.by_ref() {
+            let mut end_position = 0;
+            for (p, c) in chars.by_ref() {
                 if c == ';' {
                     is_complete = true;
+                    end_position = p + 1;
                     break;
                 }
                 entity.push(c);
             }
             if !is_complete {
-                return Err(Error::UnclosedEntity(entity));
+                return Err(ParseError::UnclosedEntity(entity, base_position + position));
             }
             change = true;
 
             if let Some(entity) = entity.strip_prefix('#') {
-                let first_char = entity
-                    .chars()
-                    .next()
-                    .ok_or_else(|| Error::InvalidEntity(entity.to_string()))?;
+                let first_char = entity.chars().next().ok_or_else(|| {
+                    ParseError::InvalidEntity(
+                        entity.to_string(),
+                        Span::new(base_position + position, base_position + end_position),
+                    )
+                })?;
                 let code = if first_char == 'x' {
                     u32::from_str_radix(&entity[1..], 16)
                 } else {
                     entity.parse::<u32>()
                 };
-                let code = code.map_err(|_| Error::InvalidEntity(entity.to_string()))?;
-                let c = std::char::from_u32(code)
-                    .ok_or_else(|| Error::InvalidEntity(entity.to_string()))?;
+                let code = code.map_err(|_| {
+                    ParseError::InvalidEntity(
+                        entity.to_string(),
+                        Span::new(base_position + position, base_position + end_position),
+                    )
+                })?;
+                let c = std::char::from_u32(code).ok_or_else(|| {
+                    ParseError::InvalidEntity(
+                        entity.to_string(),
+                        Span::new(base_position + position, base_position + end_position),
+                    )
+                })?;
                 result.push(c);
             } else {
                 match entity.as_str() {
@@ -65,7 +88,12 @@ fn parse_content(content: Cow<str>, attribute: bool) -> Result<Cow<str>, Error> 
                     "gt" => result.push('>'),
                     "lt" => result.push('<'),
                     "quot" => result.push('"'),
-                    _ => return Err(Error::InvalidEntity(entity)),
+                    _ => {
+                        return Err(ParseError::InvalidEntity(
+                            entity,
+                            Span::new(base_position + position, base_position + end_position),
+                        ))
+                    }
                 }
             }
         } else if attribute && (c == '\t' || c == '\n') {
@@ -233,21 +261,22 @@ mod tests {
     #[test]
     fn test_parse() {
         let text = "A &amp; B";
-        assert_eq!(parse_text(text.into()).unwrap(), "A & B");
+        assert_eq!(parse_text(text.into(), 0).unwrap(), "A & B");
     }
 
     #[test]
     fn test_parse_multiple() {
         let text = "&amp;&apos;&gt;&lt;&quot;";
-        assert_eq!(parse_text(text.into()).unwrap(), "&'><\"");
+        assert_eq!(parse_text(text.into(), 0).unwrap(), "&'><\"");
     }
 
     #[test]
     fn test_parse_unknown_entity() {
         let text = "&unknown;";
-        let err = parse_text(text.into());
-        if let Err(Error::InvalidEntity(entity)) = err {
+        let err = parse_text(text.into(), 0);
+        if let Err(ParseError::InvalidEntity(entity, span)) = err {
             assert_eq!(entity, "unknown");
+            assert_eq!(span, Span::new(0, 9));
         } else {
             unreachable!();
         }
@@ -256,9 +285,10 @@ mod tests {
     #[test]
     fn test_parse_unfinished_entity() {
         let text = "&amp";
-        let err = parse_text(text.into());
-        if let Err(Error::UnclosedEntity(entity)) = err {
+        let err = parse_text(text.into(), 0);
+        if let Err(ParseError::UnclosedEntity(entity, position)) = err {
             assert_eq!(entity, "amp");
+            assert_eq!(position, 0);
         } else {
             unreachable!();
         }
@@ -267,7 +297,7 @@ mod tests {
     #[test]
     fn test_parse_no_entities() {
         let text = "hello";
-        let result = parse_text(text.into()).unwrap();
+        let result = parse_text(text.into(), 0).unwrap();
         // this is the same slice
         assert!(std::ptr::eq(text, result.as_ref()));
     }
@@ -275,49 +305,49 @@ mod tests {
     #[test]
     fn test_parse_newline_r() {
         let text = "A \r B";
-        assert_eq!(parse_text(text.into()).unwrap(), "A \n B");
+        assert_eq!(parse_text(text.into(), 0).unwrap(), "A \n B");
     }
 
     #[test]
     fn test_parse_newline_rn() {
         let text = "A \r\n B";
-        assert_eq!(parse_text(text.into()).unwrap(), "A \n B");
+        assert_eq!(parse_text(text.into(), 0).unwrap(), "A \n B");
     }
 
     #[test]
     fn test_do_not_normalize_text_tab() {
         let text = "A \t B";
-        assert_eq!(parse_text(text.into()).unwrap(), "A \t B");
+        assert_eq!(parse_text(text.into(), 0).unwrap(), "A \t B");
     }
 
     #[test]
     fn test_do_not_normalize_text_newline() {
         let text = "A \n B";
-        assert_eq!(parse_text(text.into()).unwrap(), "A \n B");
+        assert_eq!(parse_text(text.into(), 0).unwrap(), "A \n B");
     }
 
     #[test]
     fn test_normalize_attribute_tab() {
         let text = "A \t B";
-        assert_eq!(parse_attribute(text.into()).unwrap(), "A   B");
+        assert_eq!(parse_attribute(text.into(), 0).unwrap(), "A   B");
     }
 
     #[test]
     fn test_normalize_attribute_r_newline() {
         let text = "A \r B";
-        assert_eq!(parse_attribute(text.into()).unwrap(), "A   B");
+        assert_eq!(parse_attribute(text.into(), 0).unwrap(), "A   B");
     }
 
     #[test]
     fn test_normalize_attribute_rn_newline() {
         let text = "A \r\n B";
-        assert_eq!(parse_attribute(text.into()).unwrap(), "A   B");
+        assert_eq!(parse_attribute(text.into(), 0).unwrap(), "A   B");
     }
 
     #[test]
     fn test_normalize_attribute_newline() {
         let text = "A \n B";
-        assert_eq!(parse_attribute(text.into()).unwrap(), "A   B");
+        assert_eq!(parse_attribute(text.into(), 0).unwrap(), "A   B");
     }
 
     #[test]
@@ -438,31 +468,31 @@ mod tests {
     #[test]
     fn test_parse_character_hex_entity() {
         let text = "A &#x26; B";
-        assert_eq!(parse_text(text.into()).unwrap(), "A & B");
+        assert_eq!(parse_text(text.into(), 0).unwrap(), "A & B");
     }
 
     #[test]
     fn test_parse_character_decimal_entity() {
         let text = "A &#38; B";
-        assert_eq!(parse_text(text.into()).unwrap(), "A & B");
+        assert_eq!(parse_text(text.into(), 0).unwrap(), "A & B");
     }
 
     #[test]
     fn test_parse_character_empty_entity() {
         let text = "A &#; B";
-        assert!(parse_text(text.into()).is_err());
+        assert!(parse_text(text.into(), 0).is_err());
     }
 
     #[test]
     fn test_parse_character_empty_hex_entity() {
         let text = "A &x#; B";
-        assert!(parse_text(text.into()).is_err());
+        assert!(parse_text(text.into(), 0).is_err());
     }
 
     #[test]
     fn test_parse_character_broken_hex_entity() {
         let text = "A &xflub#; B";
-        assert!(parse_text(text.into()).is_err());
+        assert!(parse_text(text.into(), 0).is_err());
     }
 
     #[test]