+++ src/error.rs
@@ -1,7 +1,61 @@
-use crate::xotdata::Node;
+use crate::{xotdata::Node, Span};
+
+/// An error that occurred during parsing.
+#[derive(Debug, Clone)]
+pub enum ParseError {
+    /// The XML is not well-formed - a tag is opened and never closed.
+    UnclosedTag(Span),
+    /// The XML is not well-formed - a tag is closed that was never opened.
+    InvalidCloseTag(String, String, Span),
+    /// The XML is not well-formed - you use `&` to open an entity without
+    /// closing it with `;`.
+    UnclosedEntity(String, usize),
+    /// The entity is not known. Only the basic entities are supported
+    /// right now, not any user defined ones.
+    InvalidEntity(String, Span),
+    /// You used a namespace prefix that is not declared during parsing.
+    UnknownPrefix(String, Span),
+    /// You declared an attribute of the same name twice.
+    DuplicateAttribute(String, Span),
+    /// Unsupported XML version. Only 1.0 is supported.
+    UnsupportedVersion(String, Span),
+    /// Unsupported standalone declaration. Only `yes` is supported.
+    UnsupportedNotStandalone(Span),
+    /// XML DTD is not supported.
+    DtdUnsupported(Span),
+    /// No top-level element in the document.
+    NoElementAtTopLevel(usize),
+    /// Multiple top-level elements in the document.
+    MultipleElementsAtTopLevel(Span),
+    /// Text at top level is not allowed in a well-formed document.
+    TextAtTopLevel(Span),
+    /// xmlparser error
+    XmlParser(xmlparser::Error, usize),
+}
+
+impl ParseError {
+    /// Obtain the span for a ParseError.
+    pub fn span(&self) -> Span {
+        match self {
+            ParseError::UnclosedTag(span) => *span,
+            ParseError::InvalidCloseTag(_, _, span) => *span,
+            ParseError::UnclosedEntity(_, position) => Span::new(*position, *position),
+            ParseError::InvalidEntity(_, span) => *span,
+            ParseError::UnknownPrefix(_, span) => *span,
+            ParseError::DuplicateAttribute(_, span) => *span,
+            ParseError::UnsupportedVersion(_, span) => *span,
+            ParseError::UnsupportedNotStandalone(span) => *span,
+            ParseError::DtdUnsupported(span) => *span,
+            ParseError::NoElementAtTopLevel(position) => Span::new(*position, *position),
+            ParseError::MultipleElementsAtTopLevel(span) => *span,
+            ParseError::TextAtTopLevel(span) => *span,
+            ParseError::XmlParser(_, position) => Span::new(*position, *position),
+        }
+    }
+}
 
 /// Xot errors
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub enum Error {
     // access errors
     /// The node is not a Document node.
@@ -37,29 +91,15 @@ pub enum Error {
     /// target name.
     NamespaceInProcessingInstruction,
 
-    // parser errors
-    /// The XML is not well-formed - a tag is opened and never closed.
-    UnclosedTag,
-    /// The XML is not well-formed - a tag is closed that was never opened.
-    InvalidCloseTag(String, String),
-    /// The XML is not well-formed - you use `&` to open an entity without
-    /// closing it with `;`.
-    UnclosedEntity(String),
-    /// The entity is not known. Only the basic entities are supported
-    /// right now, not any user defined ones.
-    InvalidEntity(String),
+    /// An error during parsing
+    Parse(ParseError),
+
     /// You used a namespace prefix that is not declared.
+    ///
+    /// Note that this error does not occur during parsing but during
+    /// name creation.
     UnknownPrefix(String),
-    /// You declared an attribute of the same name twice.
-    DuplicateAttribute(String),
-    /// Unsupported XML version. Only 1.0 is supported.
-    UnsupportedVersion(String),
-    /// Unsupported XML encoding. Only UTF-8 is supported.
-    UnsupportedEncoding(String),
-    /// Unsupported standalone declaration. Only `yes` is supported.
-    UnsupportedNotStandalone,
-    /// XML DTD is not supported.
-    DtdUnsupported,
+
     /// Illegal content that can never appear under a document node, such as an
     /// attribute or a namespace node
     IllegalAtTopLevel(Node),
@@ -70,13 +110,12 @@ pub enum Error {
     NoElementAtTopLevel,
     /// Multiple document elements at top level
     MultipleElementsAtTopLevel,
-    /// xmlparser error
-    Parser(xmlparser::Error),
+
     /// IO error
     ///
     /// We take the string version of the IO error so as to keep errors comparable,
     /// which is more important than the exact error object in this case (serialization)
-    Io(std::io::Error),
+    Io(String),
 }
 
 impl From<indextree::NodeError> for Error {
@@ -89,14 +128,14 @@ impl From<indextree::NodeError> for Error {
 impl From<std::io::Error> for Error {
     #[inline]
     fn from(e: std::io::Error) -> Self {
-        Error::Io(e)
+        Error::Io(e.to_string())
     }
 }
 
-impl From<xmlparser::Error> for Error {
+impl From<ParseError> for Error {
     #[inline]
-    fn from(e: xmlparser::Error) -> Self {
-        Error::Parser(e)
+    fn from(e: ParseError) -> Self {
+        Error::Parse(e)
     }
 }
 
@@ -116,22 +155,35 @@ impl std::fmt::Display for Error {
             Error::NamespaceInProcessingInstruction => {
                 write!(f, "Namespace in processing instruction target")
             }
-            Error::UnclosedTag => write!(f, "Unclosed tag"),
-            Error::InvalidCloseTag(s, s2) => write!(f, "Invalid close tag: {} {}", s, s2),
-            Error::UnclosedEntity(s) => write!(f, "Unclosed entity: {}", s),
-            Error::InvalidEntity(s) => write!(f, "Invalid entity: {}", s),
+            Error::Parse(e) => write!(f, "Parse error: {:?}", e),
             Error::UnknownPrefix(s) => write!(f, "Unknown prefix: {}", s),
-            Error::DuplicateAttribute(s) => write!(f, "Duplicate attribute: {}", s),
-            Error::UnsupportedVersion(s) => write!(f, "Unsupported version: {}", s),
-            Error::UnsupportedEncoding(s) => write!(f, "Unsupported encoding: {}", s),
-            Error::UnsupportedNotStandalone => write!(f, "Unsupported standalone"),
-            Error::DtdUnsupported => write!(f, "DTD is not supported"),
             Error::IllegalAtTopLevel(_) => write!(f, "Illegal content under document node (attribute, namespace or document node"),
             Error::TextAtTopLevel(_) => write!(f, "Text node under document not. Not allowed in a well-formed document, but allowed in a fragment"),
             Error::NoElementAtTopLevel => write!(f, "No element under document root. Not allowed in a well-formed document, but allowed in a fragment"),
             Error::MultipleElementsAtTopLevel => write!(f, "Multiple elements under document root. Not allowed in a well-formed document, but allowed in a fragment"),
-            Error::Parser(e) => write!(f, "Parser error: {}", e),
             Error::Io(s) => write!(f, "IO error: {}", s),
+        }
+    }
+}
+
+impl std::fmt::Display for ParseError {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match self {
+            ParseError::UnclosedTag(_) => write!(f, "Unclosed tag"),
+            ParseError::InvalidCloseTag(s, s2, _) => write!(f, "Invalid close tag: {} {}", s, s2),
+            ParseError::UnclosedEntity(s, _) => write!(f, "Unclosed entity: {}", s),
+            ParseError::InvalidEntity(s, _) => write!(f, "Invalid entity: {}", s),
+            ParseError::UnknownPrefix(s, _) => write!(f, "Unknown prefix: {}", s),
+            ParseError::DuplicateAttribute(s, _) => write!(f, "Duplicate attribute: {}", s),
+            ParseError::UnsupportedVersion(s, _) => write!(f, "Unsupported version: {}", s),
+            ParseError::UnsupportedNotStandalone(_) => write!(f, "Unsupported standalone"),
+            ParseError::DtdUnsupported(_) => write!(f, "DTD is not supported"),
+            ParseError::NoElementAtTopLevel(_) => write!(f, "No element at top level"),
+            ParseError::MultipleElementsAtTopLevel(_) => {
+                write!(f, "Multiple elements at top level")
+            }
+            ParseError::TextAtTopLevel(_) => write!(f, "Text at top level"),
+            ParseError::XmlParser(e, _position) => write!(f, "Parser error: {}", e),
         }
     }
 }