+++ tests/test_attribute_order.rs
@@ -1,56 +1,52 @@
-use xot::{Error, Xot};
+use xot::{ParseError, Xot};
 
 #[test]
-fn test_attribute_parser_order_is_serialization_order1() -> Result<(), Error> {
+fn test_attribute_parser_order_is_serialization_order1() {
     let mut xot = Xot::new();
     let text = r#"<doc a="A" b="B"/>"#;
-    let doc = xot.parse(text)?;
+    let doc = xot.parse(text).unwrap();
 
-    assert_eq!(xot.to_string(doc)?, text);
-    Ok(())
+    assert_eq!(xot.to_string(doc).unwrap(), text);
 }
 
 #[test]
-fn test_attribute_parser_order_is_serialization_order2() -> Result<(), Error> {
+fn test_attribute_parser_order_is_serialization_order2() {
     let mut xot = Xot::new();
     let text = r#"<doc b="B" a="A"/>"#;
-    let doc = xot.parse(text)?;
+    let doc = xot.parse(text).unwrap();
 
-    assert_eq!(xot.to_string(doc)?, text);
-    Ok(())
+    assert_eq!(xot.to_string(doc).unwrap(), text);
 }
 
 #[test]
-fn test_attribute_insert_order_is_serialization_order1() -> Result<(), Error> {
+fn test_attribute_insert_order_is_serialization_order1() {
     let mut xot = Xot::new();
     let c = xot.add_name("c");
     let text = r#"<doc a="A" b="B"/>"#;
-    let doc = xot.parse(text)?;
+    let doc = xot.parse(text).unwrap();
     let doc_el = xot.document_element(doc).unwrap();
     let mut attributes = xot.attributes_mut(doc_el);
     attributes.insert(c, "C".to_string());
-    assert_eq!(xot.to_string(doc)?, r#"<doc a="A" b="B" c="C"/>"#);
-    Ok(())
+    assert_eq!(xot.to_string(doc).unwrap(), r#"<doc a="A" b="B" c="C"/>"#);
 }
 
 #[test]
-fn test_prefix_parser_order_is_serialization_order1() -> Result<(), Error> {
+fn test_prefix_parser_order_is_serialization_order1() -> Result<(), ParseError> {
     let mut xot = Xot::new();
     let text = r#"<doc xmlns:a="A" xmlns:b="B"/>"#;
-    let doc = xot.parse(text)?;
+    let doc = xot.parse(text).unwrap();
 
-    assert_eq!(xot.to_string(doc)?, text);
+    assert_eq!(xot.to_string(doc).unwrap(), text);
     Ok(())
 }
 
 #[test]
-fn test_prefix_parser_order_is_serialization_order2() -> Result<(), Error> {
+fn test_prefix_parser_order_is_serialization_order2() {
     let mut xot = Xot::new();
     let text = r#"<doc xmlns:b="B" xmlns:a="A"/>"#;
-    let doc = xot.parse(text)?;
+    let doc = xot.parse(text).unwrap();
 
-    assert_eq!(xot.to_string(doc)?, text);
-    Ok(())
+    assert_eq!(xot.to_string(doc).unwrap(), text);
 }
 
 #[test]